cache hd seeds and ec keys

+++ src/blockchain/block/bitcoin/Bitcoin.cpp
@@ -23,7 +23,7 @@ namespace opentxs::blockchain::block::bitcoin::internal
 auto EncodeBip34(block::Height height) noexcept -> Space
 {
     if (std::numeric_limits<std::int8_t>::max() >= height) {
-        auto buf = be::little_int8_buf_t(height);
+        auto buf = be::little_int8_buf_t(static_cast<std::int8_t>(height));
         auto out = space(sizeof(buf) + 1u);
         out.at(0) = std::byte{0x1};
         std::memcpy(std::next(out.data()), &buf, sizeof(buf));
@@ -31,7 +31,7 @@ auto EncodeBip34(block::Height height) noexcept -> Space
 
         return out;
     } else if (std::numeric_limits<std::int16_t>::max() >= height) {
-        auto buf = be::little_int16_buf_t(height);
+        auto buf = be::little_int16_buf_t(static_cast<std::int16_t>(height));
         auto out = space(sizeof(buf) + 1u);
         out.at(0) = std::byte{0x2};
         std::memcpy(std::next(out.data()), &buf, sizeof(buf));
@@ -39,7 +39,7 @@ auto EncodeBip34(block::Height height) noexcept -> Space
 
         return out;
     } else if (8388607 >= height) {
-        auto buf = be::little_int32_buf_t(height);
+        auto buf = be::little_int32_buf_t(static_cast<std::int32_t>(height));
         auto out = space(sizeof(buf) + 1u);
         out.at(0) = std::byte{0x3};
         std::memcpy(std::next(out.data()), &buf, 3u);
@@ -49,7 +49,7 @@ auto EncodeBip34(block::Height height) noexcept -> Space
     } else {
         OT_ASSERT(std::numeric_limits<std::int32_t>::max() >= height);
 
-        auto buf = be::little_int32_buf_t(height);
+        auto buf = be::little_int32_buf_t(static_cast<std::int32_t>(height));
         auto out = space(sizeof(buf) + 1u);
         out.at(0) = std::byte{0x4};
         std::memcpy(std::next(out.data()), &buf, sizeof(buf));
