+++ tests/rpc/Test_Rpc.cpp
@@ -294,9 +294,9 @@ const api::Core& Test_Rpc::get_session(const std::int32_t instance)
     auto is_server = instance % 2;
 
     if (is_server) {
-        return ot::Context().Server(get_index(instance));
+        return ot::Context().Server(static_cast<int>(get_index(instance)));
     } else {
-        return ot::Context().Client(get_index(instance));
+        return ot::Context().Client(static_cast<int>(get_index(instance)));
     }
 }
